refactor: split docs graph resolver
